pyproject.toml
[project]
name = "servicewhen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["neomodel>=5.4.2", "pandas>=2.2.3", "python-dotenv>=1.0.1"]

[dependency-groups]
dev = ["ipykernel>=6.29.5", "pytest-cov>=6.0.0", "pytest>=8.3.4", "ruff>=0.8.4"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers --strict-config --cov=src --cov-fail-under=85 -ra -v"
pythonpath = "src"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "tests/**/*.py",
    "src/main.py",
    "src/**/__init__.py",
]
[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["B", "C", "D101", "D102", "D103", "F", "I", "N", "PLE", "PLW" , "PLC", "S", "UP", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]